pyproject.toml
[project]
name = "ai-steuerberater-agent"
version = "0.1.0"
description = "Decision-support AI agent for German income tax preparation (advisor-style workflow)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "google-adk",
  "fastapi>=0.110",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
